% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finplots-package.R
\docType{package}
\name{finplots-package}
\alias{finplots}
\alias{finplots-package}
\title{finplots: Financial Plots with one line of code}
\description{
Create useful visualizations for financial purposes. You only need a dataset with a date and prices, we'll take care of the rest!
}
\author{
\strong{Maintainer}: Pablo Tapia Varela \email{pablotapiav@gmail.com}

}
\keyword{internal}
