% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_boxplot.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{plot_boxplot create a boxplot}
\usage{
plot_boxplot(
  data,
  method = c("raw", "comparison"),
  frequency = c("daily", "weekly", "monthly", "yearly")
)
}
\arguments{
\item{data}{A \code{data.frame} with al least two variables.
\enumerate{
\item first variable represents dates (date format) min: daily basis (intraday not supported right now)
\item{from second variable it supports numeric values (prices)}
}}

\item{method}{A string. \code{"raw"} if you want to plot data as it comes. \code{"comparison"} transforms data on to a common base (factor 100).}

\item{frequency}{A string. Representing how do you want to see return values.}
}
\value{
a boxplot using \code{ggplot2} as base. According to your parameters will show returns on the \code{frequency} you selected.
}
\description{
plot_boxplot create a boxplot
}
\examples{
##Dummy example using base data()
airmiles_df <- data.frame(
date = as.Date(paste0(time(airmiles), "-01-01")),
airmiles = as.numeric(airmiles))

#This will generate expected plot:
plot_boxplot(airmiles_df, method = "raw", frequency = "yearly")

#For more advanced examples please visit github site.
}
