% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distr.R
\name{plot_distr}
\alias{plot_distr}
\title{plot_distr create a distribution plot}
\usage{
plot_distr(data, frequency = c("daily", "weekly", "monthly", "yearly"))
}
\arguments{
\item{data}{A \code{data.frame} with al least two variables.
\enumerate{
\item first variable represents dates (date format) min: daily basis (intraday not supported right now)
\item{from second variable it supports numeric values (prices)}
}}

\item{frequency}{A string representing how do you want to see return values}
}
\value{
a distribution plot using \link{ggplot2} as base. According to your parameters will show returns on the \code{frequency} you selected.
}
\description{
plot_distr create a distribution plot
}
\examples{
##Dummy example using base data()
airmiles_df <- data.frame(
date = as.Date(paste0(time(airmiles), "-01-01")),
airmiles = as.numeric(airmiles))
plot_distr(airmiles_df, frequency = "daily") #This will generate expected plot.

#For more advanced examples please visit github site.
}
