% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_bbg.R
\name{tidy_bbg}
\alias{tidy_bbg}
\title{tidy_bbg prepare your data coming from bbg using Rblpapi.}
\usage{
tidy_bbg(df_list)
}
\arguments{
\item{df_list}{A list. This object comes from your call to bbg. And prepare it to use it with finplots.}
}
\value{
a \code{data.frame}.
}
\description{
tidy_bbg prepare your data coming from bbg using Rblpapi.
}
\examples{
\dontrun{
#Some tickers:
names <- c("IPSA Index", "SPX Index")

Rblapi::bdh() is a function that calls data from bbg.

databbg <- bdh(
  securities = names,
  fields = "PX_LAST",
  start.date = as.Date("2013-06-01"),
  end.date = as.Date("2023-06-01")
)

#Apply function:
#dataplot <- tidy_bbg(databbg)

#Then you can use it on any finplot function:
plot_line(dataplot, method = "comparison", frequency = "daily", display = "plot")
#This will generate expected plot.

For more advanced examples please visit github site.
}
}
\seealso{
\href{https://dirk.eddelbuettel.com/code/rblpapi.html}{Rblpapi}, \code{\link[Rblpapi:bdh]{Rblpapi::bdh()}}
}
