% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_function.R
\name{tidy_data}
\alias{tidy_data}
\title{tidy_data prepare your data and makes return calcs according to your arguments given.}
\usage{
tidy_data(data, frequency = c("daily", "weekly", "monthly", "yearly"))
}
\arguments{
\item{data}{A data frame: first column must be date, from second column prices.}

\item{frequency}{You can pick daily, weekly, monthly or yearly.}
}
\description{
tidy_data prepare your data and makes return calcs according to your arguments given.
}
